WIP: storage engine refactoring #149, merge operator extracted
