+++ libakumuli/storage_engine/operators/join.cpp
@@ -114,19 +114,4 @@ std::tuple<aku_Status, size_t> JoinOperator::read(u8 *dest, size_t size) {
     return std::make_tuple(AKU_SUCCESS, output_size);
 }
 
-// Merge Join //
-
-MergeJoinOperator::MergeJoinOperator(std::vector<std::unique_ptr<TupleOperator>>&& it, bool forward)
-    : iters_(std::move(it))
-    , forward_(forward)
-{
-}
-
-std::tuple<aku_Status, size_t> MergeJoinOperator::read(u8* dest, size_t size) {
-    if (forward_) {
-        return kway_merge<0>(dest, size);
-    }
-    return kway_merge<1>(dest, size);
-}
-
 }}
