+++ unittests/CMakeLists.txt
@@ -188,6 +188,7 @@ add_executable(
     ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
     ../libakumuli/storage_engine/operators/join.cpp
+    ../libakumuli/storage_engine/operators/merge.cpp
     ../libakumuli/storage_engine/column_store.cpp
     ../libakumuli/query_processing/queryparser.cpp
     # query processor
@@ -344,6 +345,7 @@ add_executable(
     ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
     ../libakumuli/storage_engine/operators/join.cpp
+    ../libakumuli/storage_engine/operators/merge.cpp
     ../libakumuli/util.cpp
     ../libakumuli/status_util.cpp
     ../libakumuli/log_iface.cpp
@@ -373,6 +375,7 @@ add_executable(
     ../libakumuli/storage_engine/operators/aggregate.cpp
     ../libakumuli/storage_engine/operators/scan.cpp
     ../libakumuli/storage_engine/operators/join.cpp
+    ../libakumuli/storage_engine/operators/merge.cpp
     ../libakumuli/storage_engine/column_store.cpp
     ../libakumuli/util.cpp
     ../libakumuli/status_util.cpp
